Jump Capability Added
The space bar has been rigged to trigger jumping (i.e. see PlayerMovement and PlayerCollision Scripts for main changes). At this point, it is quite quirky because the foward velocity at the time of jump is not conserved. Thus, the block often slows down. After fixing this issue, it would be cool to add the slowing of the block to a fixed speed when jumping as a sort of matrix ability for the player to unlock.