Velocity in X & Z directions is now conserved when jumping
As was mentioned yesterday, the function pertaining to jumping was not adjusted to take into account the current velocity of the block when jumping. Thus, jumping did not look right. I fixed this issue in the PlayerMovement Script by retreiving the block's velocity immediately before leaving the ground via the Vector 3 class. This current velocity is inputed into the new velocity vector for the block at the time of jumping along with the desired y (i.e. upward velocity).